refactoring: extracted checkMultipleReturn, collectReferences
